Initialize repository
This commit represents the state of https://github.com/rust-bitcoin/rust-secp256k1
at commit 11e9641d21c861ef33272bdde7bf01bec73e4e7d.

We initialize this repository like this for convenience and easy of
review.

We don't want the `-zkp` variant of the bindings to be a literaly fork
because we are going to re-export most of what upstream defines.