Update to rust-secp256k1 0.20
In this version, the feature for fuzzing and external symbols have
been replaced by conditional compilation statements. As such, the
examples no longer work because we compile rustdoc with the dummy
context for fuzzing as well.

Those examples are for the upstream library anyway so we can remove
them here.