Add Tweak type
Assuming we care about constant time, we only care about constant time properties for valid keys. With the above code, we guarantee the constant time property for all valid tweaks but zero.
Although this is not ideal, this suffices in elements use-case where zero is usually as explicit commitment.

Checking this upfront might however leak information about valid Tweaks (based on the position of first 1). I can however move the code afterward with a short circuit return the valid non-zero tweak in the unsafe block.