Do docs build in Nightly job
We have a separate CI job for things that require a nightly toolchain.
Building the docs requires a nightly toolchain (because of `--cfg
docsrc` flag). It makes more sense to run the docs build in the
`Nightly` job instead of hidden in the `Tests` job.

Do the docs build in the `Nightly` job instead of in the `Tests` job.

Backport of 946ac3b51ebb8191b051c8d80bd27bee1656993e:
  https://github.com/rust-bitcoin/rust-secp256k1/pull/284 (5/9)