nits from `cargo test` and `cargo clippy` on latest nightly
The newest nightly has a new "unused mut" warning which is causing CI to
fail. Fix that, and while we're at it, run clippy. And while we're at
that, remove some `#[deny]` conditions from the crate root since these
properly belong in CI (via -D warnings) rather than in the source.