secp256k1-zkp-sys: remove not(fuzzing) gates on derives
I'm not sure why these derives are gated on not(fuzzing); I believe that
there used to be fuzzing-only manual implementations. These do not exist
anymore (or maybe they only existed upstream), and as a result the
library does not compile when fuzzing is enabled.

Also, make sure that cfg(fuzzing) is replaced by cfg(rust_secp_fuzz)
everywhere so that the user has the ability to disable fuzzing mode.