secp256k1-zkp-sys: redo patch files for latest upstream
This should really be folded into the next commit, since these patches
do not apply to the currently-vendored version of libsecp-zkp. But to
make review easier I am doing them in a separate commit.

This updates the patchfiles to delete the context and scratch space
allocation/deallocation functions. Upstream has updated the comments and
other minor stuff so the old patchfiles did not work.

While I am at it, change the patchfiles from normal diffs to git-diffs,
since we have git available and it's easier to use git since it can
figure out paths correctly.